EXTRA_DIST does not include the whole directory anymore, because that caused it to include the CVS subdirectory.  Now, only the actual data files are included in the distribution.
The DISTCLEANFILES addition satisfies the 'make distcleancheck' that is invoked at the end of 'make distcheck'.  This is not really the solution we want, because it causes the data files that are installed to be deleted.  I still need to figure out the relationship between data_DATA, $datadir, and the source, build, and install directories.


git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@45 5a19d2f1-2ada-482b-9066-51d89674df4a