After reading through the GNU Autotools documentation, I found some really good advice on how to compile subdirectories, whether they are under src/ or not.  Basically, you build a convenience library in each subdirectory, which gets LIBADD'd to the top-level surfpack library.
This has numerous advantages: No need for Makefile.am's to know
anything about sources in anything other than their own directory, no
recursive make (that is, recursion is controlled by the Autotools),
and it simplifies maintaining the build system.

To support this, I've added 'subdir-objects' to AM_INIT_AUTOMAKE, as
well as requiring version 1.9 or better of automake (to support that
option), and turned on foreign.  This last removes the requirement to
maintain two zero-byte files which were in the repository: ChangeLog
and NEWS.  I've also removed the ac_cxx_have_stl.m4 macro since we
weren't using it.


git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@278 5a19d2f1-2ada-482b-9066-51d89674df4a