Made some changes to Surfpack Interpreter syntax:
1. The type argument for CreateSurface is now unquoted and in all lowercase: kriging mars ann rbf polynomial
2. The argument named def is now axes
3. The AxesBounds command is now CreateAxes
4. GridPoints -> GridSample
5. The SaveData and SaveSurface commands have been replaced by a single Save command that takes either a surface argument or a data argument.
6. LoadSurface and LoadData have been replaced by Load (similar to 5).  The file extension determines whether the item being loaded is interpreted as data or a surface
7. All of the goodness-of-fit metrics are now unquoted
8. All of the builtin test function names are now unquoted (e.g. 'rosenbrock' -> rosenbrock)
9. The theta_vars variable for kriging is now named correlations
10. The argument named input_data is now named data
11. File extensions have been changed.  Data files must be denoted by .spd, surface files must have a .sps extension.

The system tests have been updated to reflect the new syntax.

The SurfpackInterpreter code has been overhauled, with helper functions added to to common tasks shared by all the commands.  This should make the code easier to read and maintain.

SurfpackInterface commands have been added in SurfpackInterpreter.cpp that more closely parallel (in terms of syntax and arguments) the commands used in the scripts.  The goal here is to make the use of Surfpack as a C++ library as similar to the use of Surfpack as a standalone program as similar as possible.

The mean_scaled metric ignores points with the value 'inf' (i.e., where the true function value is 0).


git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@293 5a19d2f1-2ada-482b-9066-51d89674df4a