Found and fixed a bug in the doubles_match function defined in SurfPoint.  The 'if (abs(y > 1e-10))' probe is always true as abs() works fine, but doesn't actually check whether the value of y is greater than 1e-10.  Changing the text to 'if (abs(y) > 1e-10)' recovers the original intent.
Uncovered this while changing 'using namespace std;' to multiple 'using std::*'
declarations to reduce the namespace pollution in the sources; that change will
be checked in shortly.  This fix has been unit tested on Linux and found to not
break the SurfPoint test.  TODO: abs() calls with double arguments should be
replaced with fabs().


git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@359 5a19d2f1-2ada-482b-9066-51d89674df4a