Removing the system_defs.h header in src/ann, as it duplicates the code in surfpack_system_headers.h exactly for all practical purposes; #include'ing the latter is simpler.  Also changing _all_ instances of 'using namespace std;' to however many 'using std::<STL-component>;' statements are required.
This latter usage not only provides a manifest of exactly what STL components
are used in a source file, but it also shortens the compilation time as the
whole of the std namespace isn't being brought in.  See:

http://www.gotw.ca/publications/migrating_to_namespaces.htm

I've unit tested this on Linux and found that it does _not_ break any tests.


git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@360 5a19d2f1-2ada-482b-9066-51d89674df4a