Pruned configuration parameters in preparation for team discussion. QUESTION:  how important is <pkg>_config.h as we migrate away from autotools- specific techniques for building software?
For instance, Mike and I have discovered that, if we assume modern C++
compilers, the configuration header adds little to no value for VPISparseGrid.

Assuming a migration to CMake, significant time-savings would be made by
by eliminating autotools-specific stuff.  For instance, probing for HAVE_STD
was a significant effort in CMake, so it begs the question, WHY??

Request review by any team member interested in this discussion.



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@467 5a19d2f1-2ada-482b-9066-51d89674df4a