The change in question in NKM_SurfMat.hpp is to swithc to .h less header <cassert> <cstdlib> <cstdio> as requested and also to put std:: in front of log and sqrt
The Kriging Models are works in progress.  I'm adjusting them to get better performance out of EGO, not sure what's "going wrong" with Kriging ego in  dakota_cantilever_ego.in #0 or why dakota_rosenbrock_ego.in is taking so long (about double the number of function evaluations when it finds the best (slightly better than the Dakota GP) a few samples later than the Dakota GP finds its answer.



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@576 5a19d2f1-2ada-482b-9066-51d89674df4a