Removed F77_FUNC macro redefinition warnings from the CMake build. CMake 2.8 has a new feature (FortranCInterface_HEADER) that allows the name of the F77 mangling macro to be prefixed (I chose 'SURF77_') so that it is less likely to clash with other global macros.
Using the new CMake-generated macro led to the need to touch other
files to ensure a consistent build in BOTH CMake and autotools.  I
was NOT able to figure out a way to do a similar technique with
autoheader, so the F77_FUNC macro still exists in surfpack_config.h.

Request review by Brian and/or Keith.



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@578 5a19d2f1-2ada-482b-9066-51d89674df4a