BugFix:
eig_sym() (eigen decomposition of a symmetric matrix) was copying input matrix A to result matrix eigvect and geting the LAPACK "lda" (leading dimension of A, i.e. memory footprint) from matrix A rather than eigvect.  When I wrote this it was correct, however this became a bug when I updated NKM_SurfMat to allow matrices to have different apparent (what the math says) and actual (memory footprint) sizes (the purpose of which is to reduce memory reallocation due to changing matrix size due to pivoting cholesky deciding to drop points that would make the matrix "singular").  Getting lda from A instead of eigvect was causing a segfault. 

The fix is to get lda from eigvect.




git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@607 5a19d2f1-2ada-482b-9066-51d89674df4a