NKM_SurfMat_Native.hpp: fixed an edgecase bug in uniqueelems (if all elements were unique the counter returned a value one less than the number of unique elements, I discovered it when it was caught by an assert in surfdata)
NKM_SurfMat.cpp: a few debugging statements that are #defined on or off my the surfmat error checking

NKM_ValidateMain.cpp (added some extra functions to compare Kriging and Gradient Enhanced Kriging emulators built from nested designs, and a preliminary function to generate sample designs that maximize the information in the Kriging correlation matrix)

NKM_GradKrigingModel.cpp: previously when the Kriging correlation matrix is singular gradients weren't used, now they are and additional function values are discarded to deal with conditioning but for any derivatives at a point to be used the function value at the point must also be used (i.e. a singular Kriging Correlation matrix no longer gets special treatment)

NKM_PivotChol.f: if you pass in RANK as a negative number whose absolute value is less than the number of number of rows in the matrix, this tells the routine that you only want the first abs(RANK) entries, i.e. to stop early when you are doing subset selection and you've found the number of items that you want to keep.

NKM_KrigingModel.cpp .hpp: no effect on behavior, but it is ready for a KrigingModel that uses pivoting cholesky to choose which points to use to be swapped in (that version is operational but needs further testing and a few modifications so I'm not checking it in yet)



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@608 5a19d2f1-2ada-482b-9066-51d89674df4a