found and fixed the bugs in nkm::KrigingMod::eval_gek_d2correlation_matrixdxIdJ that I had suspected were there all along
completed Mike's request, which was to use pivoted Cholesky to select a subset of trend function with the highest information content.  This resulted in somewhat unexpected behavior, it will protect against inputs with too few unique values in certain dimensions, but deliberately increasing the trend order to a "too high" value and letting pivoted cholesky select the best conditioned subset of trend functions results in Kriging Models that predict very poorly, the hand-wavey explanation for this is that trend function has stolen all of the most useful information from the Gaussian Process error model.  So the advice is use a reasonable trend order to start with and only use pivoted Cholesky of G*R^-1*G^T as insurance.

uncommented a #define statement at the top of NKM_KrigingModel.cpp to switch the Kriging Model over to the unbiased log likelihood function, this will cause a large amount of diffs in Dakota's nightly tests, some will improve slightly, some will get slightly worse but it is the theoretically right thing to do.  The reason why I hadn't done this already is that, the last time I tried this the majority of my tests got significantly worse but that no longer seems to be the case.

improved the documentation of the Kriging Model by implementing some of Brian's suggestions.



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@697 5a19d2f1-2ada-482b-9066-51d89674df4a