Apply two patches, with thanks to KRD.  These do not induce any apparent DIFFs in Dakota.
1. 5/22/2014: found a minor bug in NKM_SurfMat_Native.hpp 

In the inline member function “zero”  int nelems=getNElems(); should be int nelems=getNElemsAct();  this bug would cause an issue if the number of apparent rows is less than the number of actual rows when the zero function was called.

2.  11/24/2014: fixed a bug in the uniqueRows and uniqueCols functions

This commit also includes extensive doxygen comment additions to the SurfMat class.



git-svn-id: https://software.sandia.gov/svn/surfpack/trunk@745 5a19d2f1-2ada-482b-9066-51d89674df4a